Be/refactor/리팩터링

+++ backend/car/src/test/java/com/poongcha/car/acceptance/TrimSteps.java
@@ -63,7 +63,7 @@ public class TrimSteps {
     public static ExtractableResponse<Response> 존재하지_않는_차종으로_트림_생성_요청(
             final String trimName,
             final String imageUrl,
-            final int minPrice,
+            final long minPrice,
             final long carTypeId
     ) {
         return given()
